html/infrastructure/safe-passing-of-structured-data/shared-array-buffers/nested-inheriting-worker-success.https.html

<!doctype html>
<meta charset=utf-8>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<div id=log></div>
<script>
const url = new URL("./", self.location);

const workerScript = `
importScripts("${url.origin}/resources/testharness.js");
importScripts("${url.href}resources/test-incrementer.js");

promise_test(t => {
  const worker = new Worker("${url.href}resources/incrementer-worker.js");

  return testSharingViaIncrementerScript(t, worker, "parent worker", worker, "sub-worker");
}, "postMessaging to a dedicated sub-worker allows them to see each others' modifications");

test(() => {
  assert_true(self.crossOriginIsolated);
}, "Bonus: self.crossOriginIsolated");

done();
`;

fetch_tests_from_worker(new Worker(URL.createObjectURL(new Blob([workerScript], { type: "text/javascript" }))));

// First test times out?
// data: shouldn't be cross-origin isolated, maybe?
// fetch_tests_from_worker(new Worker(`data:,${workerScript}`));

</script>
